fix: move validation service into utils
